refactor: useScroll hook 생성
