+++ client/src/hooks/Main/useScroll.ts
@@ -1,0 +1,40 @@
+import { getAllPosts } from '@/service/get-posts';
+import { useInfiniteQuery } from '@tanstack/react-query';
+import { useEffect } from 'react';
+
+function useScroll() {
+  const { data, fetchNextPage, hasNextPage, isFetching } = useInfiniteQuery({
+    queryKey: ['projects'],
+    queryFn: getAllPosts,
+    getNextPageParam: (lastPage) => lastPage?.nextPageParms,
+    suspense: true,
+  });
+
+  useEffect(() => {
+    const position = Number(sessionStorage.getItem('scrollPosition')) || 0;
+    if (position > 0) {
+      window.scrollTo(0, position);
+      sessionStorage.setItem('scrollPosition', '0');
+    }
+  }, []);
+
+  useEffect(() => {
+    const handleScroll = async (event: Event) => {
+      if (event.target instanceof Document) {
+        const { scrollHeight, scrollTop, clientHeight } = event.target
+          .scrollingElement as Element;
+        if (!isFetching && scrollHeight - scrollTop <= clientHeight * 1.2) {
+          await fetchNextPage();
+        }
+      }
+    };
+    document.addEventListener('scroll', handleScroll);
+    return () => {
+      document.removeEventListener('scroll', handleScroll);
+    };
+  }, [hasNextPage, isFetching]);
+
+  return { datas: data?.pages, hasNextPage };
+}
+
+export default useScroll;
